+++ src/Application/AddAsset.php
@@ -83,6 +83,7 @@ class AddAsset
         }
 
         if (is_string($file)) {
+            if(!$filename) $filename = md5(time()) . '.'. substr($file, 11, strpos($file, ';') - 11);
             return AssetUploader::uploadFromBase64($file, $filename);
         }